pull out terms and sources
