+++ polar-c-api/src/lib.rs
@@ -1,5 +1,5 @@
 pub use polar_core::polar::{Polar, Query};
-use polar_core::{error, types};
+use polar_core::{error, types, terms};
 
 use std::cell::RefCell;
 use std::ffi::{CStr, CString};
@@ -115,7 +115,7 @@ pub extern "C" fn polar_register_constant(
         let value = serde_json::from_str(&value);
         match value {
             Ok(value) => {
-                polar.register_constant(types::Symbol::new(name.as_ref()), value);
+                polar.register_constant(terms::Symbol::new(name.as_ref()), value);
                 POLAR_SUCCESS
             }
             Err(e) => {
@@ -237,8 +237,8 @@ pub extern "C" fn polar_debug_command(query_ptr: *mut Query, value: *const c_cha
         if !value.is_null() {
             let s = unsafe { ffi_string!(value) };
             let t = serde_json::from_str(&s);
-            match t.as_ref().map(types::Term::value) {
-                Ok(types::Value::String(command)) => match query.debug_command(command) {
+            match t.as_ref().map(terms::Term::value) {
+                Ok(terms::Value::String(command)) => match query.debug_command(command) {
                     Ok(_) => POLAR_SUCCESS,
                     Err(e) => {
                         set_error(e);
