+++ polar-core/src/error.rs
@@ -2,6 +2,8 @@ use serde::{Deserialize, Serialize};
 
 use std::fmt;
 
+use crate::sources::*;
+use crate::terms::*;
 use crate::types::*;
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
