+++ polar-core/src/lib.rs
@@ -13,6 +13,8 @@ mod numerics;
 pub mod parser;
 pub mod polar;
 mod rewrites;
+mod sources;
+pub mod terms;
 pub mod types;
 mod vm;
 mod warnings;
